requirements.txt
requests
python-dotenv
alpaca-py
alpaca-trade-api
yfinance
torch
torchaudio
torchvision
transformers
lumibot
APScheduler
numpy
pandas